Move all MetalsPlugin code back into one file
